:recycle: refactor(payment): Create package of response request dto
