+++ src/main/java/nbc/mushroom/domain/payment/dto/request/PaymentReq.java
@@ -1,4 +1,4 @@
-package nbc.mushroom.domain.payment.dto;
+package nbc.mushroom.domain.payment.dto.request;
 
 public record PaymentReq(
     String paymentKey,
